cmd/vfsgendev: Split package into multiple .go files.
Feedback suggests this increases readability. It feels about the same
for me, but I hope it helps other people who are less familiar with the
code, since I'd rather optimize for other people's readability (I'm
already very familiar with it).
